pyproject.toml
[project]
name = "django-landing"
version = "0.1.0"
description = ""
authors = [
    {name = "sergey",email = "sbitza@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "django-mptt (>=0.18.0,<0.19.0)",
    "django-extensions (>=4.1,<5.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "environs (>=14.3.0,<15.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-admin-sortable2 (>=2.2.8,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "cssutils (>=2.11.1,<3.0.0)",
    "lxml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
faker = "^37.8.0"

